requirements.txt
Flask
pycryptodome
gunicorn
python-dotenv
psycopg2